Pull pass
Some editorial changes, nits and clarifications, but I also removed a bit of text that was copied from RFC7341 again (this is just not needed) and re-added some minor parts from the previous draft version. I also re-added a deployment consideration section as the client not being aware is a mayor difference to RFC7341 and needs some care in deployment. This is a bit redundant with some part of the sec considerations now but I think that is fine as this point is important.

In general, I think this proposed change makes the draft shorter and more to the point.